working

+++ src/app/service.py
@@ -48,8 +48,8 @@ class Service:
             query = HistoricalWinLossQuery(request_obj=request)
             job_config = bigquery.QueryJobConfig(
                 query_parameters=[
-                    bigquery.ScalarQueryParameter("team1_code", "INT", request.team1_code.ncaa_code),
-                     bigquery.ScalarQueryParameter("team2_code", "INT", request.team2_code.ncaa_code)
+                    bigquery.ScalarQueryParameter("team1_code", "INT64", request.team1_code.ncaa_code),
+                     bigquery.ScalarQueryParameter("team2_code", "INT64", request.team2_code.ncaa_code)
                 ]
             )
             result = self.client.execute_query(query=query.get_query(), job_config=job_config)
